Add sample rate to codec profiles
This fixes issues with converting from DSD source files. A better solution
would be to check the existing files' sample rates and compare against a
"maximum" setting to see if we want to resample the audio. e.g. if maximum
sample rate is set to 48 kHz then anything above that will be resampled
to 48 kHz, anything at or below 48 kHz will not be resampled.